refactor(backoffice-v2): improved typography and spacing in the table cell
